requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
Pillow>=10.0.0
flask>=3.0.0
flask-login>=0.6.3
flask-sqlalchemy>=3.1.0
werkzeug>=3.0.0

